+++ generator/python/templates/sdk.tmpl
